pyproject.toml
[project]
name = "nanoRecSys"
version = "0.1.0"
authors = [
    {name = "Zhe Liu"}
]
description = "A nano recommendation system project"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "faiss-cpu",
    "ipykernel",
    "matplotlib",
    "mypy",
    "nbstripout",
    "numpy",
    "pandas",
    "pyarrow",
    "pydantic-settings >=2.0.0",
    "pytest",
    "pytorch-lightning",
    "redis",
    "requests",
    "ruff",
    "torch",
    "torchmetrics",
    "tqdm",
    "uvicorn",
    "wandb",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "nbstripout",
    "locust",
    "streamlit",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["nanoRecSys*"]
